requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
sentence-transformers
transformers
python-dotenv
pypdf
requests
streamlit
openai
